backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
openai
pytesseract
Pillow
python-dotenv
numpy
opencv-python
pydub
aiofiles